% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ultility.R
\name{simulate_adjacency}
\alias{simulate_adjacency}
\title{Simulate an adjacency matrix}
\usage{
simulate_adjacency(n, K, alpha, pi, directed = FALSE)
}
\arguments{
\item{n}{an integer: the number of nodes}

\item{K}{an integer: the number of cluster}

\item{alpha}{a matrix:}

\item{pi}{a vector of size K: the mixture parameter}

\item{directed}{Wether the graph is directed or not}
}
\value{
A n*n adjacency matrix
}
\description{
Simulate an adjacency matrix
}
